+++ packages/cluster/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/cluster
 
+## 0.5.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.6.0
+
 ## 0.5.3
 
 ### Patch Changes