+++ packages/config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/config
 
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.6.0
+
 ## 0.2.4
 
 ### Patch Changes