+++ packages/ipc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shardix/ipc
 
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.6.0
+
 ## 0.4.4
 
 ### Patch Changes