+++ packages/provider-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shardix/provider-config
 
+## 0.3.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.6.0
+  - @shardix/config@0.2.5
+
 ## 0.3.4
 
 ### Patch Changes