+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shardix/testing
 
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @shardix/common@0.6.0
+  - @shardix/core@0.6.0
+
 ## 0.7.1
 
 ### Patch Changes